refactor: finish structure
