+++ src/services/users/createUser.js
@@ -1,6 +1,6 @@
 const { v4: uuidv4 } = require('uuid');
-const { database } = require('../repository/database');
-const { User } = require('../models/User');
+const { database } = require('../../repository/database');
+const { User } = require('../../models/User');
 
 const createUser = ({ data }) => {
   try {
