+++ src/services/users/readUsers.js
@@ -1,4 +1,4 @@
-const { database } = require("../repository/database");
+const { database } = require("../../repository/database");
 
 const readUsers = () => {
   try {
